ARM: EXYNOS: Check overflow & Remove useless code
If total is bigger than ccnt, ccnt is set to total

+++ arch/arm/mach-exynos/busfreq_opp_4x12.c
@@ -383,7 +383,7 @@ int exynos4x12_init(struct device *dev, struct busfreq_data *data)
 {
 	unsigned int i;
 	unsigned int tmp;
-	unsigned long maxfreq = UINT_MAX;
+	unsigned long maxfreq = ULONG_MAX;
 	unsigned long minfreq = 0;
 	int ret;
